% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_formula.R
\name{add_covars_linear}
\alias{add_covars_linear}
\title{Title}
\usage{
add_covars_linear(x, specifications)
}
\arguments{
\item{x}{an expression}

\item{specifications}{a list of symbols}
}
\value{
an expression x
#' @importFrom rlang expr
}
\description{
Title
}
