% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_formula.R
\name{add_covars_nonlinear}
\alias{add_covars_nonlinear}
\title{Title}
\usage{
add_covars_nonlinear(x, nl_formulae, specifications)
}
\arguments{
\item{x}{an expression}

\item{nl_formulae}{a list of expressions}

\item{specifications}{a list of symbols}
}
\value{
list of expressions
}
\description{
Title
}
