% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_formula.R
\name{add_skill_terms_1PL}
\alias{add_skill_terms_1PL}
\title{Title}
\usage{
add_skill_terms_1PL(x, nl_formulae, var_specs, add_common_dimension)
}
\arguments{
\item{add_common_dimension}{}
}
\value{

}
\description{
Title
}
