% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{aggregate_warnings}
\alias{aggregate_warnings}
\title{Aggregates identical warnings
Used to get a warning only once when it occures more often (e.g. from an expression in a loop or purrr::map())
Code inside expression will be executed and therefore alters / creates variables in parent environment.}
\usage{
aggregate_warnings(expr)
}
\arguments{
\item{expr}{expression to execute}
}
\description{
Aggregates identical warnings
Used to get a warning only once when it occures more often (e.g. from an expression in a loop or purrr::map())
Code inside expression will be executed and therefore alters / creates variables in parent environment.
}
