% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birtm_aio.R
\name{birtm_aio}
\alias{birtm_aio}
\title{All-in-one bayesian item response theory modeler}
\usage{
birtm_aio(
  response_data,
  response_columns,
  person_data = NULL,
  item_data = NULL,
  situation_data = NULL,
  variable_specifications = NULL,
  model_specifications = NULL,
  prior = NULL,
  cores = 4,
  file = NULL,
  check_prior_influence = FALSE,
  preparing_bf = FALSE,
  add_waic = FALSE,
  add_loo = FALSE,
  refit = FALSE,
  ...
)
}
\arguments{
\item{response_data}{A tibble or dataframe with person identifiers and responsdata (and possibly person covariables).}

\item{response_columns}{A character or symbol vector with all columns that represent item responses or a range of columns specified corresponding to dplyr::select() rules (e. g.: item01:item25).}

\item{person_data}{A tibble or dataframe with person identifiers linked to person covariables.}

\item{item_data}{A tibble or dataframe with item identifiers linked to item covariables.}

\item{situation_data}{A tibble or dataframe with person and item identifiers linked to situation covariables.}

\item{variable_specifications}{Named list of characters or strings.}

\item{model_specifications}{Named list of strings and numerics.}

\item{prior}{Bayesian priors in format of brms::prior}

\item{cores}{Number of CPU cores used to fit the model}

\item{file}{Path and name the fit object is saved for later reusage (will be loaded automatically if already exists)}

\item{check_prior_influence}{Boolean}

\item{preparing_bf}{Booleans}

\item{add_waic}{Boolean}

\item{add_loo}{Boolean}

\item{refit}{Boolean}

\item{...}{Arguments passed to brms::brm}
}
\value{
birtsmsfit object
}
\description{
A wrapper function to fit a bayesian IRT model with one command.
}
\examples{
\dontrun{
testdata <- data_spm[1:100,] # take only 1/5th of the data to speed up the examplary model fitting
fit_1d_1pl_spm2 <- birtms::birtm_aio(response_data = testdata, response_columns = i1:i12,
file = '../inst/extdata/fit_1d_1pl_spm2')
fit_1d_1pl_spm2
}
}
