% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_formula.R
\name{build_formula}
\alias{build_formula}
\title{Builds a brmsformula}
\usage{
build_formula(variable_specifications = NULL, model_specifications = NULL)
}
\arguments{
\item{variable_specifications}{Named list of characters or strings.}

\item{model_specifications}{Named list of strings and numerics.}
}
\value{
Returns an object of type brmsformula to use it with brms::brm()
}
\description{
Builds a brmsformula for an IRT model based on user input. This is the main function for formula generantion.
It calls further functions to build an appropriate formula. It can be called without it's arguments which
loads the standard settings for a linear 1PL model.
}
\examples{
build_formula()
}
\seealso{
\code{\link[brms]{brmsformula}}
}
