% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_formula.R
\name{build_formula_nonlinear}
\alias{build_formula_nonlinear}
\title{Title}
\usage{
build_formula_nonlinear(
  var_specs,
  add_common_dimension = FALSE,
  item_parameter_number
)
}
\arguments{
\item{var_specs}{a named list of symbols}

\item{add_common_dimension}{boolean}

\item{item_parameter_number}{numeric (more precise integer)}
}
\value{
brmsformula
}
\description{
Title
}
