% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_latent.R
\name{calc_latent_regression_coefs_distribution}
\alias{calc_latent_regression_coefs_distribution}
\title{Calculates distribution of latent regression coefficients}
\usage{
calc_latent_regression_coefs_distribution(variance, beta, Y)
}
\arguments{
\item{variance}{dataframe of variances. Column for each dimension. Row per posteriorsample draw.}

\item{beta}{dataframe of regressioncoefficients. Column for each predictor. Row per posteriorsample draw.}

\item{Y}{dataframe of regression indicators. Column for each predictor. Row per observation.}
}
\value{
List with R.squared and sd median_hdi summary.
}
\description{
Calculates distribution of latent regression coefficients
}
