% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_latent.R
\name{calc_latent_regression_coefs_point_estimator}
\alias{calc_latent_regression_coefs_point_estimator}
\title{Calculates point estimate of latent regression coefficients}
\usage{
calc_latent_regression_coefs_point_estimator(variance, beta, Y)
}
\arguments{
\item{variance}{numeric value}

\item{beta}{numeric vector}

\item{Y}{numeric matrix}
}
\value{
List of R.quared and sd value.
}
\description{
Calculates point estimate of latent regression coefficients
}
