% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{calculate_odds_ratio}
\alias{calculate_odds_ratio}
\title{Odds ratio
Calculates the odds ratio for the posterior samples or the original responses
adjusted from Anna Scharl and Timo Gnambs (https://www.tqmp.org/RegularArticles/vol15-2/p075/p075.pdf)}
\usage{
calculate_odds_ratio(y_rep = NULL, y = NULL, zero_correction = FALSE)
}
\arguments{
\item{y_rep}{(pers) x (item) x (rep) array; replicated data (can handle response data as dataframe or lists of dataframes as well)}

\item{y}{(pers) x (item) dataframe; response data}

\item{zero_correction}{boolean; should Haldane zero-correction be used? (Add 0.5 to all cells?)}
}
\value{
tibble
}
\description{
Odds ratio
Calculates the odds ratio for the posterior samples or the original responses
adjusted from Anna Scharl and Timo Gnambs (https://www.tqmp.org/RegularArticles/vol15-2/p075/p075.pdf)
}
\examples{
tibble::tribble(
~x, ~y,
0, 1,
1, 0,
1, 1,
0, 0,
1, 0
) \%>\% calculate_odds_ratio() # equals 0.5
}
