% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{calculate_odds_ratio_parallel}
\alias{calculate_odds_ratio_parallel}
\title{Odds ratio parallel
Calculates the odds ratio for the posterior samples or the original responses
adjusted from Anna Scharl and Timo Gnambs (https://www.tqmp.org/RegularArticles/vol15-2/p075/p075.pdf)}
\usage{
calculate_odds_ratio_parallel(
  y_rep = NULL,
  y = NULL,
  zero_correction = "none",
  ci_method = "Woolf",
  ci_width = 0.89,
  nsim_ci = 1e+07,
  nsim_median = 1e+05,
  cores = 4
)
}
\arguments{
\item{y_rep}{(pers) x (item) x (rep) array; replicated data (can handle response data as dataframe or lists of dataframes as well)}

\item{y}{(pers) x (item) dataframe; response data}

\item{zero_correction}{character; 'none', 'Haldane', or 'Bayes'}

\item{ci_method}{character; 'Woolf', 'unconditional', 'compromise', 'BayesEqTails' or 'BayesHDI'}

\item{ci_width}{double}

\item{nsim_ci}{integer; number of draws used to get a distribution of the odds ratio value to calculate CIs}

\item{nsim_median}{integer; number of draws used to get a distribution of the odds ratio value to calculate the median}

\item{cores}{int; number of CPU cores used for parallel processing}
}
\value{
tibble
}
\description{
During Haldane zero-correction a value close to 0 is added to all cells
(https://www.oxfordreference.com/view/10.1093/acref/9780199976720.001.0001/acref-9780199976720-e-1977).
Most common is 0.5. This value eliminates the first order bias term.
}
\details{
Bayesian functions not checked for correct implementation with parallel function
}
\examples{
tibble::tribble(
~x, ~y,
0, 1,
1, 0,
1, 1,
0, 0,
1, 0
) \%>\% calculate_odds_ratio() # equals 0.5
}
