% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_helpers.R
\name{check_and_set_specifications}
\alias{check_and_set_specifications}
\title{Checks model specifications and possibly adds standard settings}
\usage{
check_and_set_specifications(specifications)
}
\arguments{
\item{specifications}{a named list of characters}
}
\value{
a list
}
\description{
Calls override_standard_specifications() to override the standard
specifications with user specifications. Checks if all specifications (the
names of the list) come from a list of valid names.
}
