% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_loglik.R
\name{check_n_nodes}
\alias{check_n_nodes}
\title{Checks marginal loglik for different numbers of nodes}
\usage{
check_n_nodes(fit, min_nodes = 1, max_nodes = 8, cores = 4)
}
\arguments{
\item{fit}{birtmsfit object}

\item{min_nodes}{int; minimal nodenumber is 2^(min_nodes+1)+1}

\item{max_nodes}{int; maximum nodenumber is 2^(max_nodes+1)+1}

\item{cores}{int; number of CPU cores used for parallel processing}
}
\value{
list to use with birtms::plot_check_n_nodes
}
\description{
Checks marginal loglik for different numbers of nodes
}
