% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{contingency2successratio}
\alias{contingency2successratio}
\title{Calculate successes per sample and sample sizes from contingency table}
\usage{
contingency2successratio(mat)
}
\arguments{
\item{mat}{vector of length 4 with counts c(n11, n00, n10, n01) from contingency table}
}
\value{
vector of length 4: c(y1, n1, y2, n2)
}
\description{
Calculate successes per sample and sample sizes from contingency table
}
