% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{count_for_itempair_or}
\alias{count_for_itempair_or}
\title{Creates contingency table counts for item pair based odds ratio}
\usage{
count_for_itempair_or(x, y)
}
\arguments{
\item{x}{double vector, answers to item 1}

\item{y}{double vector, answers to item 2}
}
\value{
double vector of length 4: c(n11, n00, n10, n01)
}
\description{
Creates contingency table counts for item pair based odds ratio
}
