% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{custom_skim}
\alias{custom_skim}
\title{Skim table}
\usage{
custom_skim(data, ci_width = 0.89)
}
\arguments{
\item{data}{data.frame}

\item{ci_width}{double}
}
\value{
data.table
}
\description{
Summarises multiple columns of a table. Works with grouping.
}
