% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_spm}
\alias{data_spm}
\title{Evaluated answers for 499 persons of the SPM-LS}
\format{
A data frame with 499 rows and 13 variables:
\describe{
  \item{person}{person id}
  \item{i1}{answers to item 1}
  \item{i2}{answers to item 2}
  \item{i3}{answers to item 3}
  \item{i4}{answers to item 4}
  \item{i5}{answers to item 5}
  \item{i6}{answers to item 6}
  \item{i7}{answers to item 7}
  \item{i8}{answers to item 8}
  \item{i9}{answers to item 9}
  \item{i10}{answers to item 10}
  \item{i11}{answers to item 11}
  \item{i12}{answers to item 12}
  ...
}
}
\source{
\url{https://data.mendeley.com/datasets/h3yhs5gy3w/1}
}
\usage{
data_spm
}
\description{
A dataset containing the binary evaluated answers of 499 persons to the 12 hardest items of the SPM-LS collected by Myszkowski and Storme (2018).
}
\keyword{datasets}
