% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_interval.R
\name{dens_sj_bounded}
\alias{dens_sj_bounded}
\title{Density with SJ-bandwidth method also suitable for bounded densities}
\usage{
dens_sj_bounded(dat, n = 512, cut = TRUE, na.rm = FALSE)
}
\arguments{
\item{dat}{numeric vector}

\item{n}{integer, number of points where density gets evaluated}

\item{cut}{boolean, should only be FALSE for internal use}

\item{na.rm}{boolean}
}
\value{
S3 density
}
\description{
Density with SJ-bandwidth method also suitable for bounded densities
}
