% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_mode}
\alias{get_mode}
\title{Get mode
Code from: https://www.tutorialspoint.com/r/r_mean_median_mode.htm}
\usage{
get_mode(v)
}
\arguments{
\item{v}{vector}
}
\value{
vector of length 1
}
\description{
Get mode
Code from: https://www.tutorialspoint.com/r/r_mean_median_mode.htm
}
\examples{
v <- c(2, 2, 1)
get_mode(v)
}
