% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{get_or}
\alias{get_or}
\title{Summarises Odds Ratio statistic
Returns odds ratio values for actual dataset and posterior predictions.
Summarises the mode and hdi of their difference and returns the ppp-value.}
\usage{
get_or(model, n_samples = NULL, hdi_width = 0.89, zero_correction = FALSE)
}
\arguments{
\item{model}{birtmsfit}

\item{n_samples}{int - Number of posterior smaples to use}

\item{hdi_width}{double}

\item{zero_correction}{boolean; should Haldane zero-correction be used? (Add 0.5 to all cells?)}
}
\value{
birtmsdata; tibble with additinal attributes
}
\description{
Summarises Odds Ratio statistic
Returns odds ratio values for actual dataset and posterior predictions.
Summarises the mode and hdi of their difference and returns the ppp-value.
}
\examples{
\dontrun{
get_or(fit, n_samples = 500)
}
}
