% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{get_or_distribution}
\alias{get_or_distribution}
\title{Get odds ratio distribution}
\usage{
get_or_distribution(counts, k = 0.5, nsim = 1e+07)
}
\arguments{
\item{counts}{(4x1) matrix with counts n11, n00, n10, n01 from contingency table}

\item{k}{double; concentration of beta priors: 0.5 for Jeffreys prior, 1 for uniform priors}

\item{nsim}{interger}
}
\value{
double, odds ratio distribution
}
\description{
Get odds ratio distribution
}
\examples{
c <- c(5,2,10,7) # n11, n00, n10, n01
or_dist <- get_or_distribution(c, nsim = 10)
}
