% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.R
\name{get_postdata}
\alias{get_postdata}
\title{Generates a list of posterior predictions and probabilities}
\usage{
get_postdata(model = NULL, subset = NULL)
}
\arguments{
\item{model}{birtmsfit object}

\item{subset}{integer vector; which posterior samples should be used?}
}
\value{
list of data.frames
}
\description{
Generates a list of posterior predictions and probabilities
}
