% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPMC.R
\name{get_ppmccriteria}
\alias{get_ppmccriteria}
\title{Calculates posterior predictive check for specific criteria}
\usage{
get_ppmccriteria(model, ppmcdata, criteria, ppmcMethod, group = "item")
}
\arguments{
\item{model}{birtmsfit}

\item{ppmcdata}{data.frame; data generated with birtms::get_ppmcdatasets}

\item{criteria}{character; currently 'infit', 'outfit' or 'll'}

\item{ppmcMethod}{char; either 'C' for consevative checks or 'M' for mixed ppmc with independently drawn theta distribution}

\item{group}{char; 'item' or column name of person identifier (e.g. 'person', 'id', 'token')}
}
\value{
data.frame to use with birtms::plot_fit_statistic
}
\description{
Calculates posterior predictive check for specific criteria
}
