% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPMC.R
\name{get_ppmcdatasets}
\alias{get_ppmcdatasets}
\title{Generates data for posterior predictive model checks}
\usage{
get_ppmcdatasets(
  model,
  ppmcMethod,
  post_responses = NULL,
  sd = 1,
  n_samples = NULL,
  sequential = FALSE
)
}
\arguments{
\item{model}{birtmsfit object}

\item{ppmcMethod}{char; either 'C' for consevative checks or 'M' for mixed ppmc with independently drawn theta distribution}

\item{post_responses}{data.frame; data generated with birtms::get_postdata}

\item{sd}{double; standard deviation prior for theta distrubution (may be adjusted if person covars are used)}

\item{n_samples}{integer; number of posterior samples to use (using all posterior samples might lead to memory overflow)}

\item{sequential}{boolean; should mixed PPMC data be calculated sequential (slower but less memory intensive); try adjusting n_samples first}
}
\value{
data.frame to use with birtms::get_ppmccriteria
}
\description{
Generates data for posterior predictive model checks
}
