% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hsm_hdi}
\alias{hsm_hdi}
\title{Mode via hsm and HDI
define a point_interval function using the hsm (half sample mode) estimator from modeest}
\usage{
hsm_hdi(...)
}
\arguments{
\item{...}{see ggdist::point_interval()}
}
\value{
dataframe
}
\description{
Mode via hsm and HDI
define a point_interval function using the hsm (half sample mode) estimator from modeest
}
\examples{
hsm_hdi(c(1,2,3,3,2.5))
}
