% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_interval.R
\name{hsm_hdi_sjb}
\alias{hsm_hdi_sjb}
\title{Mode via hsm and HDI also for bounded densities
define a point_interval function using the hsm (half sample mode) estimator from modeest}
\usage{
hsm_hdi_sjb(
  .data,
  ...,
  .width = 0.89,
  inf.rm = FALSE,
  na.rm = FALSE,
  tie.limit = 0.05
)
}
\arguments{
\item{.data}{dataframe}

\item{...}{columns to get point_interval for}

\item{.width}{double, ci width}

\item{inf.rm}{boolean, should infinite and NaN values be dropped}

\item{na.rm}{boolean, should NA values be dropped}

\item{tie.limit}{double}
}
\value{
dataframe
}
\description{
Mode via hsm and HDI also for bounded densities
define a point_interval function using the hsm (half sample mode) estimator from modeest
}
\examples{
hsm_hdi_sjb(c(1,2,3,3,2.5))
}
