% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.R
\name{ICC_check}
\alias{ICC_check}
\title{Plots ICC for an item}
\usage{
ICC_check(
  model,
  item_id = 1,
  num_groups = NULL,
  verbose = FALSE,
  post_responses = NULL,
  ellipse_type = "t",
  plot_post_person_estimated = TRUE
)
}
\arguments{
\item{model}{birtmsfit object}

\item{item_id}{int; item to plot ICC for}

\item{num_groups}{int; number of groups of persons}

\item{verbose}{boolean; should a explanation be added as a caption}

\item{post_responses}{list of data.frames from birtms::get_postdata}

\item{ellipse_type}{string; 'axisparallel' or 'norm' or 't' or 'none'}

\item{plot_post_person_estimated}{boolean}
}
\value{
ggplot object
}
\description{
Plots ICC for an item
}
