% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_loglik.R
\name{loo_marginal}
\alias{loo_marginal}
\title{Caclulates marginal loo}
\usage{
loo_marginal(fit, ...)
}
\arguments{
\item{fit}{birtmsfit object}

\item{...}{parameters passed to birms::marginal_loglik}
}
\value{
loo
}
\description{
Caclulates marginal loo
}
