% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_responsedata_wider}
\alias{make_responsedata_wider}
\title{Make response data wider
Transforms dataset from a brmsfit from long into wide format}
\usage{
make_responsedata_wider(model)
}
\arguments{
\item{model}{brmsfit object}
}
\value{
tibble
}
\description{
Make response data wider
Transforms dataset from a brmsfit from long into wide format
}
\examples{
\dontrun{
make_responsedata_wider(fit)
}
}
