% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_loglik.R
\name{marginal_loglik}
\alias{marginal_loglik}
\title{marginal_loglik
Calculates the marginal loglikelihood for an IRT model}
\usage{
marginal_loglik(fit, n_nodes = 11, cores = 4)
}
\arguments{
\item{fit}{birtmsfit object}

\item{n_nodes}{int; number of nodes where the loglik gets evaluated; some models need more nodes to reach convergence of loglik}

\item{cores}{int; number of CPU cores used for parallel processing}
}
\value{
list
}
\description{
marginal_loglik
Calculates the marginal loglikelihood for an IRT model
}
