% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_interval.R
\name{median_hdi_sjb}
\alias{median_hdi_sjb}
\title{Median and HDI also for bounded densities}
\usage{
median_hdi_sjb(.data, ..., .width = 0.89, inf.rm = FALSE, na.rm = FALSE)
}
\arguments{
\item{.data}{dataframe}

\item{...}{columns to get point_interval for}

\item{.width}{double, ci width}

\item{inf.rm}{boolean, should infinite and NaN values be dropped}

\item{na.rm}{boolean, should NA values be dropped}
}
\value{
dataframe
}
\description{
Median and HDI also for bounded densities
}
\examples{
#' median_hdi_sjb(c(1,2,3,3,2.5))
}
