% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birtmsfit.R
\name{new_birtmsfit}
\alias{new_birtmsfit}
\title{Created a birtmsfit from brms fit by adding specs}
\usage{
new_birtmsfit(
  fit,
  variable_specifications = NULL,
  model_specifications = NULL,
  file = NULL
)
}
\arguments{
\item{fit}{brmsfit object}

\item{variable_specifications}{list of stringt}

\item{model_specifications}{list of stringt}

\item{file}{character; filepath}
}
\value{
birtmsfit object
}
\description{
Created a birtmsfit from brms fit by adding specs
}
