% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{or.sim}
\alias{or.sim}
\title{Simulate Odds Ratio distribution
Extracted from PropCIs::orci.bayes()}
\usage{
or.sim(a1, b1, c1, d1, nsim = 1e+07)
}
\arguments{
\item{a1}{a parameter for first beta distribution}

\item{b1}{b parameter for first beta distribution}

\item{c1}{a parameter for second beta distribution}

\item{d1}{b parameter for second beta distribution}

\item{nsim}{integer}
}
\value{
vector of doubles, distribution od odds ratios
}
\description{
Simulate Odds Ratio distribution
Extracted from PropCIs::orci.bayes()
}
