% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{or_ci_bayes}
\alias{or_ci_bayes}
\title{Calculates CI for Odds Ratio bayesian way
Extracted from PropCIs::orci.bayes()
Can return HDI instead of equal tailed CI interval as well.}
\usage{
or_ci_bayes(counts, conf.level = 0.89, k = 0.5, nsim = 1e+07, hdi = FALSE)
}
\arguments{
\item{counts}{double of length 4}

\item{conf.level}{double}

\item{k}{double; prior for beta distribution}

\item{nsim}{integer}

\item{hdi}{boolean; should HDI be used instead of equaly tailes CI?}
}
\value{
double of length 2
}
\description{
Calculates CI for Odds Ratio bayesian way
Extracted from PropCIs::orci.bayes()
Can return HDI instead of equal tailed CI interval as well.
}
