% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{or_ci_uncond}
\alias{or_ci_uncond}
\title{Wrapper to get bayesian CI for OR}
\usage{
or_ci_uncond(counts, ci = 0.89)
}
\arguments{
\item{counts}{double of length 4}

\item{ci}{double}
}
\value{
double of length 2
}
\description{
Wrapper to get bayesian CI for OR
}
