% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{or_ci_woolf}
\alias{or_ci_woolf}
\title{Calculates CIs for Odds ratio with Woolfs method}
\usage{
or_ci_woolf(or, counts, corr = 0.5, ci_width = 0.89)
}
\arguments{
\item{or}{double}

\item{counts}{double of length 4}

\item{corr}{double; Haldane zero correction term}

\item{ci_width}{double}
}
\value{
double of length 2
}
\description{
Calculates CIs for Odds ratio with Woolfs method
}
