% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{or_distribution_bayes}
\alias{or_distribution_bayes}
\title{Get the odds ratio distribution
combines prior with actual data to posterior beta distributions and initiates sampling
Extracted from PropCIs::orci.bayes()}
\usage{
or_distribution_bayes(x1, n1, x2, n2, a, b, c, d, nsim = 1e+07)
}
\arguments{
\item{x1}{number of successes in sample 1}

\item{n1}{sample size in sample 1}

\item{x2}{number of successes in sample 2}

\item{n2}{sample size in sample 2}

\item{a}{beta prior for x1}

\item{b}{beta prior for x2}

\item{c}{beta prior for n1}

\item{d}{beta prior for n2}

\item{nsim}{integer}
}
\value{
vector of doubles, distribution od odds ratios
}
\description{
Get the odds ratio distribution
combines prior with actual data to posterior beta distributions and initiates sampling
Extracted from PropCIs::orci.bayes()
}
