% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{or_limits_irt}
\alias{or_limits_irt}
\title{Calculate Odds Ratio Limits for IRT models
Calculates limits for odds ratio values for itempairs to check if they show
misfit assuming that data is generated by a one-dimensional 1PL or 2PL model.
LIMITS ON LOG ODDS RATIOS FOR UNIDIMENSIONAL ITEM RESPONSE THEORY MODELS (Haberman, 2007)}
\usage{
or_limits_irt(ai = 1, aj = 1, sigma = 1, beta = 0)
}
\arguments{
\item{ai}{double; slope of item 1}

\item{aj}{double; slope of item 2}

\item{sigma}{double; standard deviation}

\item{beta}{double}
}
\value{
list(lower_limit, upper_limit) of doubles
}
\description{
Calculate Odds Ratio Limits for IRT models
Calculates limits for odds ratio values for itempairs to check if they show
misfit assuming that data is generated by a one-dimensional 1PL or 2PL model.
LIMITS ON LOG ODDS RATIOS FOR UNIDIMENSIONAL ITEM RESPONSE THEORY MODELS (Haberman, 2007)
}
