% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{or_median_bayes}
\alias{or_median_bayes}
\title{Get median of odds ratio distribution}
\usage{
or_median_bayes(mat, k = 0.5, nsim = 1e+07)
}
\arguments{
\item{mat}{(4x1) matrix with counts n11, n00, n10, n01 from contingency table}

\item{k}{double; concentration of beta priors: 0.5 for Jeffreys prior, 1 for uniform priors}

\item{nsim}{interger}
}
\value{
median od odds ratio distribution
}
\description{
Get median of odds ratio distribution
}
