% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{or_mode_bayes}
\alias{or_mode_bayes}
\title{Get mode of odds ratio distribution}
\usage{
or_mode_bayes(counts, k = 0.5, nsim = 1e+07)
}
\arguments{
\item{counts}{(4x1) matrix with counts n11, n00, n10, n01 from contingency table}

\item{k}{double; concentration of beta priors: 0.5 for Jeffreys prior, 1 for uniform priors}

\item{nsim}{interger}
}
\value{
mode of odds ratio distribution
}
\description{
Get mode of odds ratio distribution
}
