% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation_helpers.R
\name{override_standard_specifications}
\alias{override_standard_specifications}
\title{Augmants the standard specifications with user specifications}
\usage{
override_standard_specifications(specifications, type)
}
\arguments{
\item{specifications}{a named list}

\item{type}{the type of the}
}
\value{
a named list
}
\description{
Augmants the standard specifications with user specifications
}
