% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_loglik.R
\name{plot_check_n_nodes}
\alias{plot_check_n_nodes}
\title{Plots the marginal loglik and number of problematic pareto k values for different numbers of nodes}
\usage{
plot_check_n_nodes(check_n_nodes_object)
}
\arguments{
\item{check_n_nodes_object}{list from birtms::check_n_nodes}
}
\value{
ggplot object
}
\description{
Plots the marginal loglik and number of problematic pareto k values for different numbers of nodes
}
