% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPMC.R
\name{plot_fit_statistic}
\alias{plot_fit_statistic}
\title{Densityplots of posterior predicitve checks}
\usage{
plot_fit_statistic(
  model,
  data,
  units = c(1, 9),
  group = "item",
  ppmcMethod = "C",
  hdi_width = 0.89
)
}
\arguments{
\item{model}{birtmsfit object}

\item{data}{data.frame; data generated with birtms::get_ppmccriteria}

\item{units}{integer vector of length 2; first and last item/person to generate a plot for}

\item{group}{char; 'item' or column name of person identifier (e.g. 'person', 'id', 'token')}

\item{ppmcMethod}{char; either 'C' for consevative checks or 'M' for mixed ppmc with independently drawn theta distribution}

\item{hdi_width}{double; sets width of credibility interval}
}
\value{
ggplot object
}
\description{
Densityplots of posterior predicitve checks
}
