% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_itempars.R
\name{plot_itemparameter}
\alias{plot_itemparameter}
\title{Plots cateye or dot plots for item parameters}
\usage{
plot_itemparameter(
  model,
  pars,
  style,
  items = c(1, 5),
  thresholds = c(0.2, 2),
  alphacut = c(0.2, 0.3, 2),
  betacut = c(-2, 2)
)
}
\arguments{
\item{model}{birtmsfit object}

\item{pars}{character; either 'slope', 'easyness', 'delta', 'beta', 'difficulty' or 'pseudoguess'}

\item{style}{character; either "halfeye" or "dots"}

\item{items}{integer vector of length 2; indicates first and last item to plot the parameter distributions for}

\item{thresholds}{double vector of length 2; crops the plot horizontally}

\item{alphacut}{double vector of length 3;}

\item{betacut}{double vector of length 2;}
}
\value{
ggplot object
}
\description{
Plots cateye or dot plots for item parameters
}
