% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personresponsefunction.R
\name{plot_personresponsefunction}
\alias{plot_personresponsefunction}
\title{PLots person response functions}
\usage{
plot_personresponsefunction(model, personresponsedata, id, persResFunStyle = 1)
}
\arguments{
\item{model}{birtmsfit object}

\item{personresponsedata}{data.frame from birtms::calc.personresponsedata}

\item{id}{integer vector; persons to plot PRF for}

\item{persResFunStyle}{iteger vector; 1 = scatter, 2 = posterior predict based PRF, 3 = pp_scatter, 4 = force_se}
}
\value{
ggplot object
}
\description{
PLots person response functions
}
