% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ppmc_ridges.R
\name{plot_ppmc_distribution}
\alias{plot_ppmc_distribution}
\title{Plot PPMC parameter distributions}
\usage{
plot_ppmc_distribution(
  data,
  column = NULL,
  method = "SJexpanded",
  ci_width = 0.89,
  density_ci = FALSE,
  smooth_density_ci = FALSE,
  color = "lightblue",
  n = 128,
  clean_data = FALSE
)
}
\arguments{
\item{data}{numeric vector or dataframe}

\item{column}{if data is a datframe specify a single column that should a plot be returned for}

\item{method}{character; 'Mueller94' or 'SJexpanded'}

\item{ci_width}{double}

\item{density_ci}{boolean; should CI for density should be bootstrapped?}

\item{smooth_density_ci}{boolean; should CI be smoothed?}

\item{color}{color}

\item{n}{integer, number of points of desity estimation; can be smaller for Mueller94 (start with 128)}

\item{clean_data}{boolean; should infinite values and NaNs get removed?}
}
\value{
ggplot2 object
}
\description{
Plot PPMC parameter distributions
}
\examples{
\dontrun{
or_data <- get_or(fit, n_samples = 500)
or_data \%>\% select(-or_rep_samples) \%>\% filter(item1 == 1, item2 == 2) \%>\%
birtsms::unnest_keep_attr(or_dif_samples) \%>\% select(or_dif)
birtsms::plot_ppmc_distribution()
}
}
