% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ppmc_ridges.R
\name{plot_ppmc_ridges}
\alias{plot_ppmc_ridges}
\title{Plot PPMC parameter distributions}
\usage{
plot_ppmc_ridges(
  data,
  parameter,
  group = 0,
  rope = FALSE,
  hdi = TRUE,
  ci_width = 0.89,
  color = "lightblue",
  range = NULL,
  custom_ci = NULL,
  ...
)
}
\arguments{
\item{data}{tibble}

\item{parameter}{column name that holds numeric values to plot density ridges for}

\item{group}{column name to optionally specify that multiple density ridges should be plotted regarding the grouping value}

\item{rope}{boolean; should the rope be plotted as a gray area?}

\item{hdi}{boolean; should HDI or equitailed region be plotted}

\item{ci_width}{double}

\item{color}{color; sets color of ridges}

\item{range}{double; c(min, max) used for setting xlim in coord_cartesian()}

\item{custom_ci}{double vector to pass the ci limits manually (e.g. when )}

\item{...}{arguments passed to ggridges::stat_density_ridges()}
}
\value{
ggplot2 object
}
\description{
Plot PPMC parameter distributions
}
\examples{
\dontrun{
or_data <- get_or(fit, n_samples = 500)
or_data \%>\% select(-or_rep_samples) \%>\% filter(item1 == 1, item2 == 2) \%>\%
birtsms::unnest_keep_attr(or_dif_samples) \%>\%
birtsms::plot_ppmc_distribution(or_dif)

or_data \%>\% select(-or_rep_samples) \%>\% filter(item1 == 1, item2 \%in\% 2:4) \%>\%
birtsms::unnest_keep_attr(or_dif_samples) \%>\%
birtsms::plot_ppmc_distribution(or_dif, group = item2, rope = TRUE)
}
}
