% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrightmap.R
\name{plot_wrightmap}
\alias{plot_wrightmap}
\title{Plots a wrightmap}
\usage{
plot_wrightmap(
  model,
  items = c(1:5),
  palette = NULL,
  lims = NULL,
  namefun = identity,
  groupfun = identity,
  labsize = 4,
  bins = 20,
  classic = TRUE
)
}
\arguments{
\item{model}{birtmsfit object}

\item{items}{integer vector; which items should be mapped?}

\item{palette}{color palette}

\item{lims}{double vector of length 2}

\item{namefun}{function to shorten names}

\item{groupfun}{function to group items based on names}

\item{labsize}{double}

\item{bins}{integer}

\item{classic}{boolean}
}
\value{
ggplot object
}
\description{
Plots a wrightmap
}
