% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{posterior_epred_long}
\alias{posterior_epred_long}
\title{Expected Values of the Posterior Predictive Distribution in long format}
\usage{
posterior_epred_long(model, n_samples = NULL)
}
\arguments{
\item{model}{brmsfit}

\item{n_samples}{integer}
}
\value{
tibble with columns from brmsfit$data (at least original responses, person and item identifier).
Additionally ppe (posterior predictive estimate) and .draw (which MCMC draw comes the answer from).
}
\description{
Expected Values of the Posterior Predictive Distribution in long format
}
\examples{
\dontrun{
posterior_epred_long(fit, n_samples = 500)
}
}
