% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{posterior_predict_long}
\alias{posterior_predict_long}
\title{Samples from the Posterior Predictive Distribution in long format}
\usage{
posterior_predict_long(model, n_samples = NULL)
}
\arguments{
\item{model}{brmsfit}

\item{n_samples}{integer}
}
\value{
tibble with columns from brmsfit$data (at least original responses, person and item identifier).
Additionally yrep (predicted answers) and .draw (which MCMC draw comes the answer from).
}
\description{
Samples from the Posterior Predictive Distribution in long format
}
