% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{posterior_predictive_values_long}
\alias{posterior_predictive_values_long}
\title{Returns posterior predictive values in long format}
\usage{
posterior_predictive_values_long(model, n_samples = NULL, f)
}
\arguments{
\item{model}{brmsfit}

\item{n_samples}{integer}

\item{f}{function}
}
\value{
tibble with columns from brmsfit$data (at least original responses, person and item identifier).
Additionally y_rep (predicted answers) or ppe (posterior predictive estimate) and .draw (which MCMC draw comes the answer from).
}
\description{
Returns posterior predictive values in long format
}
