% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birtsmsdata.R
\name{print.birtmsdata}
\alias{print.birtmsdata}
\title{Print birtms data
Prints data as a tibble but prints additional attributes as well}
\usage{
\method{print}{birtmsdata}(x, ...)
}
\arguments{
\item{x}{birtmsdata}

\item{...}{see base::print()}
}
\description{
Print birtms data
Prints data as a tibble but prints additional attributes as well
}
\examples{
\dontrun{
or_data <- get_or(fit, n_samples = 500)
or_data
}
}
