% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_latent.R
\name{R2_latent}
\alias{R2_latent}
\title{R.squared}
\usage{
R2_latent(birtms_fit, fast = TRUE)
}
\arguments{
\item{birtms_fit}{Object of type birtmsfit.}

\item{fast}{Boolean. If true observations get summarised by persons or items. Attention: It's not tested if the results are the same for datasets with missings or when situationcovars are added!}
}
\value{
List with R.squared and sd value for each dimension for persons and items.
}
\description{
Computes R.squared for 1PL models with item and person covariates. (Hierarchical or higher parametric models not tested right now.)
}
