% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_formula.R
\name{set_person_grouping}
\alias{set_person_grouping}
\title{Title}
\usage{
set_person_grouping(var_specs)
}
\arguments{
\item{var_specs}{a named list of symbols}
}
\value{
an expression p
}
\description{
Title
}
