% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{timeit}
\alias{timeit}
\title{Takes time of a process using tictoc package}
\usage{
timeit(expr)
}
\arguments{
\item{expr}{expression in curled brackets}
}
\value{
result of expression
}
\description{
Takes time of a process using tictoc package
}
