% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unnest_keep_attr}
\alias{unnest_keep_attr}
\title{Unnesting conserving outer attributes}
\usage{
unnest_keep_attr(...)
}
\arguments{
\item{...}{see tidyr::unnest()}
}
\value{
tibble
}
\description{
Unnesting conserving outer attributes
}
\examples{
\dontrun{
data <- get_or(fit, n_samples = 500)
unnest_keep_attr(data) \%>\% attr('rope_width') # conserves rope_width attribute
unnest(data) \%>\% attr('rope_width') # drops attribute
}
}
