% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_formula.R
\name{add_skill_terms_2PL}
\alias{add_skill_terms_2PL}
\title{Title}
\usage{
add_skill_terms_2PL(x, nl_formulae, var_specs, add_common_dimension)
}
\arguments{
\item{x}{an expression}

\item{nl_formulae}{a list of expressions}

\item{var_specs}{a named list of symbols}

\item{add_common_dimension}{boolean}
}
\value{
list of expressions
}
\description{
Title
}
