% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birtm.R
\name{birtm}
\alias{birtm}
\title{Bayesian item response theory modeler}
\usage{
birtm(
  data,
  formula,
  prior = NULL,
  cores = 4,
  file = NULL,
  check_prior_influence = FALSE,
  preparing_bf = FALSE,
  add_waic = TRUE,
  add_loo = TRUE,
  refit = FALSE,
  variable_specifications = NULL,
  model_specifications = NULL,
  ...
)
}
\arguments{
\item{data}{Data in long format (i.e. composed with birtms::compose_data)}

\item{formula}{A brms::formula (i.e. builded with birtms::build_formula)}

\item{prior}{Bayesian priors in format of brms::prior}

\item{cores}{Number of CPU cores used to fit the model}

\item{file}{Path and name the fit object is saved for later reusage (will be loaded automatically if already exists)}

\item{check_prior_influence}{Boolean}

\item{preparing_bf}{Boolean}

\item{add_waic}{Boolean}

\item{add_loo}{Boolean}

\item{refit}{Boolean}

\item{variable_specifications}{A named list}

\item{model_specifications}{A named list}

\item{...}{Arguments passed to brms::brm}
}
\value{
birtsmsfit object
}
\description{
Fits a bayesian IRT model.
}
\examples{
\dontrun{
testdata <- birtms::compose_dataset(response_data = data_spm[1:100,], response_columns = i1:i12)
formula_1d_1pl <- birtms::build_formula()
fit_1d_1pl_spm1 <- birtms::birtm(data = testdata, formula = formula_1d_1pl,
file = '../inst/extdata/fit_1d_1pl_spm1')
fit_1d_1pl_spm1
}
}
