% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_formula.R
\name{build_formula_linear}
\alias{build_formula_linear}
\title{Title}
\usage{
build_formula_linear(var_specs, add_common_dimension = FALSE)
}
\arguments{
\item{add_common_dimension}{}
}
\value{

}
\description{
Title
}
