% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_formula.R
\name{build_formula_nonlinear}
\alias{build_formula_nonlinear}
\title{Title}
\usage{
build_formula_nonlinear(var_specs, mod_specs)
}
\arguments{
\item{var_specs}{a named list of symbols}

\item{mod_specs}{Named list of strings and numerics.}
}
\value{
brmsformula
}
\description{
Title
}
