% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_interval.R
\name{get_mode}
\alias{get_mode}
\title{Get mode}
\usage{
get_mode(v, ...)
}
\arguments{
\item{v}{vector}

\item{...}{arguments passed to modeest::hsm()}
}
\value{
vector
}
\description{
Get mode
}
\examples{
v <- c(2, 2, 1)
get_mode(v)
}
