% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{get_or}
\alias{get_or}
\title{Summarises Odds Ratio statistic
Returns odds ratio values for actual dataset and posterior predictions.
Summarises the mode and hdi of their difference and returns the ppp-value.}
\usage{
get_or(
  model,
  n_samples = NULL,
  ci_width = 0.89,
  zero_correction = "none",
  ci_method = "Woolf",
  nsim_act = 1e+07,
  nsim_rep = 1e+05,
  tie.limit = 0.05
)
}
\arguments{
\item{model}{birtmsfit}

\item{n_samples}{int - Number of posterior smaples to use}

\item{ci_width}{double}

\item{zero_correction}{character; 'none', 'Haldane', 'compromise' or 'Bayes'}

\item{ci_method}{character; 'Woolf', 'unconditional', 'compromise', 'BayesEqTails' or 'BayesHDI'}

\item{nsim_act}{integer}

\item{nsim_rep}{integer}

\item{tie.limit}{double}
}
\value{
birtmsdata; tibble with additinal attributes
}
\description{
Summarises Odds Ratio statistic
Returns odds ratio values for actual dataset and posterior predictions.
Summarises the mode and hdi of their difference and returns the ppp-value.
}
\examples{
\dontrun{
get_or(fit, n_samples = 500)
}
}
