% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hsm_hdi}
\alias{hsm_hdi}
\title{Mode via hsm and HDI
define a point_interval function using the hsm (half sample mode) estimator from modeest}
\usage{
hsm_hdi(
  .data,
  ...,
  .width = 0.89,
  inf.rm = FALSE,
  nan.rm = FALSE,
  tie.limit = 0.05
)
}
\arguments{
\item{.data}{dataframe}

\item{...}{columns to get point_interval for}

\item{.width}{double, ci width}

\item{inf.rm}{boolean, should infinite values be dropped}

\item{nan.rm}{boolean, should NaN values be dropped}

\item{tie.limit}{double}
}
\value{
dataframe
}
\description{
Mode via hsm and HDI
define a point_interval function using the hsm (half sample mode) estimator from modeest
}
\examples{
hsm_hdi(c(1,2,3,3,2.5))
}
