% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_loglik.R
\name{marginal_loglik}
\alias{marginal_loglik}
\title{marginal_loglik
Calculates the marginal loglikelihood for an IRT model}
\usage{
marginal_loglik(fit, n_nodes = 11, cores = 4)
}
\arguments{
\item{cores}{}
}
\value{

}
\description{
marginal_loglik
Calculates the marginal loglikelihood for an IRT model
}
