% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{plot_or_heatmap}
\alias{plot_or_heatmap}
\title{Odds ratio heatmap
Odds ratio values should be greater 1 for items that are conditionally independent in a one dimensional
IRT model (Holland, Rosenbaum, 1986). Haberman (2007) shows that the lower bond can be estimated more
precisely (and will be even higher than 1) and also a upper bond can be calculated and gives formulae
for 1PL and 2PL models.}
\usage{
plot_or_heatmap(
  or_data,
  model = NULL,
  a = 1,
  sigma = 1,
  median_centered_colorscale = TRUE,
  bayesian = FALSE,
  nsim_ci = 1e+06,
  ci_width = 0.89
)
}
\arguments{
\item{or_data}{dataframe generated by birtms::get_or()}

\item{model}{birtmsfit object}

\item{a}{double; mean slope parameter (if not passed model)}

\item{sigma}{double; standard deviation of person dimesnion (theta)}

\item{median_centered_colorscale}{boolean}

\item{bayesian}{boolean; should the bayesian distribution for actual odds ratio should be used as reference instead of point estimate}

\item{nsim_ci}{integer; number of draws used to get a distribution of the odds ratio value to calculate CIs}

\item{ci_width}{double}
}
\value{
ggplot2 object
}
\description{
Odds ratio heatmap
Odds ratio values should be greater 1 for items that are conditionally independent in a one dimensional
IRT model (Holland, Rosenbaum, 1986). Haberman (2007) shows that the lower bond can be estimated more
precisely (and will be even higher than 1) and also a upper bond can be calculated and gives formulae
for 1PL and 2PL models.
}
\examples{
\dontrun{
or_data <- get_or(fit, n_samples = 500)
plot_or_heatmap(or_data)
}
}
