% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{plot_ppmc_or_heatmap}
\alias{plot_ppmc_or_heatmap}
\title{PPMC Odds ratio heatmap
Prints a heatmap of odds ratio differences used for post predictive model checking.
Be aware that setting use_rope to TRUE will highlight fewer items as probably problematic.}
\usage{
plot_ppmc_or_heatmap(or_data, use_rope = FALSE, alternative_color = FALSE)
}
\arguments{
\item{or_data}{dataframe generated by birtms::get_or()}

\item{use_rope}{boolean; not only check if 0 is included in HDI but rather any value of a ROPE}

\item{alternative_color}{boolean; color highlight only the problematic items}
}
\value{
ggplot object
}
\description{
PPMC Odds ratio heatmap
Prints a heatmap of odds ratio differences used for post predictive model checking.
Be aware that setting use_rope to TRUE will highlight fewer items as probably problematic.
}
\examples{
\dontrun{
or_data <- get_or(fit, n_samples = 500)
plot_ppmc_or_heatmap(or_data)
}
}
