% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2_latent.R
\name{R2_latent}
\alias{R2_latent}
\title{Bayesian kind of R.squared}
\usage{
R2_latent(birtms_fit, fast = TRUE)
}
\arguments{
\item{birtms_fit}{Object of type birtmsfit.}

\item{fast}{Boolean. If true observations get summarised by persons or items. Should work also for datasets with missings and different observation numbers per person. Attention: It's not tested if the results are the same for datasets when situationcovars are added!}
}
\value{
List with R.squared and sd value for each dimension for persons and items.
}
\description{
Computes R.squared for non hierarchical models with item and person covariates. (Hierarchical models not tested right now.)
}
